Text summarizer Chatbot/requirements.txt
langchain==0.2.14
langchain_google_genai==1.0.9
python-dotenv==1.0.1
streamlit==1.31.0
streamlit==1.37.1
